Skip complete feed downloads
Compare chunks of RSS feed requests with previous cached values
and abort download if the tail of the downloaded bytes will be already
found in the cached value.

This is problematic if older entries of a feed will be changed. The
complete download of the feed can still be forced by using
the cache=0 URI argument.