Add cache cleaning and feed._uid
Before this commit a feed was identified by its name.
This was replaced by an id (hash value, no collision detection)
This id is used as cache filename, too.

The memory and disk usage of the server can now be limited by
settings.CACHE_MEMORY_LIMIT and settings.CACHE_DISK_LIMIT.

Currently, this limiting just counts the bytes in the cache, but not the
overall memory usage. Later commits should respect the (big) feed.context
variables.