Use multiprocess.Pool for actions
Replacing the fork/double_fork approach by using Python's
multiprocessing package.
The new class ActionPool will handle the processes.
Currently, only a fixed duration for all actions is defined.
After this period an action-process is abort able (if the user requests a new
one).

(User defined) actions (settings.ACTIONS) now returning
a) A function handle. Then, this function will be called in the
   main thread. This could blocking the applications.

   or

b) A PickableAction-object. This will provided to a worker process
   of the ActionPool.